refactor: 소셜 로그인 로직 개선
