+++ tamago-core/src/main/kotlin/tamago/server/core/user/UserFacade.kt
@@ -7,6 +7,9 @@ import tamago.server.core.refreshtoken.RefreshTokenCommandUseCase
 import tamago.server.core.refreshtoken.RefreshTokenQueryUseCase
 import tamago.server.core.user.application.exception.EmailAlreadyExistsException
 import tamago.server.core.user.application.exception.InvalidCredentialsException
+import tamago.server.core.user.application.exception.UserSaveErrorException
+import tamago.server.core.user.application.service.UserCommandService
+import tamago.server.core.user.application.service.UserQueryService
 import tamago.server.core.user.domain.port.inbound.command.LoginCommandDto
 import tamago.server.core.user.domain.port.inbound.command.SignUpCommandDto
 import tamago.server.core.user.domain.port.inbound.command.TestLoginCommandDto
@@ -14,8 +17,8 @@ import tamago.server.core.user.domain.port.inbound.query.TokenQueryDto
 
 @Component
 class UserFacade(
-    private val userCommandUseCase: UserCommandUseCase,
-    private val userQueryUseCase: UserQueryUseCase,
+    private val userCommandService: UserCommandService,
+    private val userQueryService: UserQueryService,
     private val refreshTokenCommandUseCase: RefreshTokenCommandUseCase,
     private val refreshTokenQueryUseCase: RefreshTokenQueryUseCase,
     private val jwtTokenProvider: JwtTokenProvider,
@@ -23,27 +26,45 @@ class UserFacade(
 ) {
     fun signUp(command: SignUpCommandDto) =
         command
-            .takeUnless { userQueryUseCase.exists(it.email) }
-            ?.let(userCommandUseCase::createUser)
+            .takeUnless { userQueryService.exists(it.email) }
+            ?.let(userCommandService::createUser)
             ?: throw EmailAlreadyExistsException()
 
-    fun socialLogin(command: LoginCommandDto): TokenQueryDto = userCommandUseCase.socialLogin(command)
+    fun socialLogin(command: LoginCommandDto): TokenQueryDto {
+        val user =
+            userQueryService.findByExternalId(command.provider, command.externalId)
+                ?: userCommandService.createSocialUser(command)
+
+        val userId = user.id ?: throw UserSaveErrorException()
+
+        val accessToken = jwtTokenProvider.generateAccessToken(userId, user.role.name)
+        val refreshToken = jwtTokenProvider.generateRefreshToken(userId, user.role.name)
+
+        refreshTokenCommandUseCase.saveOrUpdate(userId, refreshToken)
+
+        return TokenQueryDto(
+            userId = userId.value,
+            accessToken = accessToken,
+            refreshToken = refreshToken,
+            isNewUser = !user.isOnboarded(),
+        )
+    }
 
     fun emailLogin(command: TestLoginCommandDto): TokenQueryDto {
-        val auth = userQueryUseCase.getEmailAuth(command.email)
+        val auth = userQueryService.getEmailAuth(command.email)
 
         require(passwordEncoder.matches(command.password, auth.password)) {
             throw InvalidCredentialsException()
         }
 
         val userId = auth.userId ?: throw InvalidCredentialsException()
-        val user = userQueryUseCase.get(userId)
+        val user = userQueryService.get(userId)
 
         val accessToken = jwtTokenProvider.generateAccessToken(userId, user.role.name)
         val refreshToken = jwtTokenProvider.generateRefreshToken(userId, user.role.name)
 
         refreshTokenCommandUseCase.saveOrUpdate(userId, refreshToken)
-        userCommandUseCase.recordLogin(user)
+        userCommandService.recordLogin(user)
 
         return TokenQueryDto(
             userId = userId.value,
@@ -55,7 +76,7 @@ class UserFacade(
 
     fun reissueToken(refreshToken: String): TokenQueryDto {
         val userId = jwtTokenProvider.getUserId(refreshToken)
-        val user = userQueryUseCase.get(userId)
+        val user = userQueryService.get(userId)
 
         refreshTokenQueryUseCase.validation(userId, refreshToken)
