+++ tamago-core/src/main/kotlin/tamago/server/core/user/application/service/UserCommandService.kt
@@ -6,24 +6,19 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import tamago.server.core.common.event.UserDeletedEvent
 import tamago.server.core.common.event.UserSignedUpEvent
-import tamago.server.core.common.jwt.JwtTokenProvider
 import tamago.server.core.common.vo.UserId
-import tamago.server.core.refreshtoken.RefreshTokenCommandUseCase
 import tamago.server.core.user.UserCommandUseCase
 import tamago.server.core.user.application.exception.UserSaveErrorException
 import tamago.server.core.user.domain.aggregate.User
 import tamago.server.core.user.domain.enum.AuthProvider
 import tamago.server.core.user.domain.port.inbound.command.LoginCommandDto
 import tamago.server.core.user.domain.port.inbound.command.SignUpCommandDto
-import tamago.server.core.user.domain.port.inbound.query.TokenQueryDto
 import tamago.server.core.user.domain.port.outbound.UserPersistencePort
 
 @Service
 @Transactional
 class UserCommandService(
     private val userPersistencePort: UserPersistencePort,
-    private val refreshTokenCommandUseCase: RefreshTokenCommandUseCase,
-    private val jwtTokenProvider: JwtTokenProvider,
     private val passwordEncoder: PasswordEncoder,
     private val applicationEventPublisher: ApplicationEventPublisher,
 ) : UserCommandUseCase {
@@ -42,25 +37,10 @@ class UserCommandService(
         publishUserSignedUpEvent(userId)
     }
 
-    override fun socialLogin(command: LoginCommandDto): TokenQueryDto {
-        val user =
-            userPersistencePort.findByExternalId(command.provider, command.externalId)
-                ?: run { createSocialUser(command) }
-
-        val userId = user.id ?: throw UserSaveErrorException()
-
-        val accessToken = jwtTokenProvider.generateAccessToken(userId, user.role.name)
-        val refreshToken = jwtTokenProvider.generateRefreshToken(userId, user.role.name)
-
-        refreshTokenCommandUseCase.saveOrUpdate(userId, refreshToken)
-
-        return TokenQueryDto(
-            userId = userId!!.value,
-            accessToken = accessToken,
-            refreshToken = refreshToken,
-            isNewUser = !user.isOnboarded(),
-        )
-    }
+    fun createSocialUser(command: LoginCommandDto) =
+        userPersistencePort
+            .save(User.create(command.email, command.provider, command.externalId))
+            .also { publishUserSignedUpEvent(it.id!!) }
 
     override fun updateNickname(
         user: User,
@@ -87,14 +67,6 @@ class UserCommandService(
     ) {
         user.addRunningDistance(distance)
         userPersistencePort.save(user)
-    }
-
-    private fun createSocialUser(command: LoginCommandDto): User {
-        val savedUser = userPersistencePort.save(User.create(command.email, command.provider, command.externalId))
-        val userId = savedUser.id ?: throw UserSaveErrorException()
-
-        publishUserSignedUpEvent(userId)
-        return savedUser
     }
 
     private fun publishUserSignedUpEvent(userId: UserId) {
